+++ src/com/reason/comp/dune/DuneProcess.java
@@ -5,9 +5,9 @@ import com.intellij.execution.configurations.*;
 import com.intellij.execution.process.*;
 import com.intellij.openapi.project.*;
 import com.intellij.openapi.vfs.*;
-import com.reason.comp.Compiler;
+import com.reason.comp.*;
 import com.reason.*;
-import com.reason.ide.console.*;
+import com.reason.comp.Compiler;
 import org.jetbrains.annotations.*;
 
 import java.util.*;