+++ src/com/reason/comp/esy/EsyProcess.java
@@ -5,11 +5,11 @@ import com.intellij.execution.configurations.*;
 import com.intellij.execution.process.*;
 import com.intellij.openapi.project.*;
 import com.intellij.openapi.vfs.*;
-import com.reason.comp.Compiler;
+import com.reason.comp.*;
 import com.reason.*;
+import com.reason.comp.Compiler;
 import com.reason.comp.bs.*;
 import com.reason.ide.*;
-import com.reason.ide.console.*;
 import org.jetbrains.annotations.*;
 
 import java.util.*;