+++ src/com/reason/ide/console/bs/BsMakeAction.java
@@ -1,21 +1,23 @@
-package com.reason.ide.console;
+package com.reason.ide.console.bs;
 
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.project.Project;
+import com.reason.comp.*;
+import com.reason.ide.console.*;
 import org.jetbrains.annotations.NotNull;
 
-public class DuneBuildAction extends CompilerAction {
+public class BsMakeAction extends CompilerAction {
 
-  public DuneBuildAction() {
-    super("Build dune", "Build dune", AllIcons.Actions.Compile);
+  public BsMakeAction() {
+    super("Make bsb", "Make bsb", AllIcons.Actions.Compile);
   }
 
   @Override
   public void actionPerformed(@NotNull AnActionEvent e) {
     Project project = e.getProject();
     if (project != null) {
-      doAction(project, CliType.Dune.BUILD);
+      doAction(project, CliType.Bs.MAKE);
     }
   }
 }